refactor(resource): split manager.rs into 7 submodules (Tech Spec §5.4)
Pure structural move. NO public API change — all imports continue
to resolve at nebula_resource::*.

Breakdown:
- manager/mod.rs       — Manager struct + register*/acquire* surface
- manager/options.rs   — ManagerConfig, RegisterOptions, ShutdownConfig,
                         DrainTimeoutPolicy
- manager/registration.rs — register_inner + reverse-index write
- manager/gate.rs      — GateAdmission + admit_through_gate +
                         settle_gate_admission (#322)
- manager/execute.rs   — resilience pipeline (execute_with_resilience,
                         validate_pool_config)
- manager/rotation.rs  — ResourceDispatcher trampoline + on_credential_*
                         (moved from crates/resource/src/rotation.rs)
- manager/shutdown.rs  — graceful_shutdown + drain helpers +
                         set_phase_all + set_phase_all_failed (#302, R-023)

Field visibility relaxed to pub(super) where cross-submodule access
required — sibling modules under manager/ now access fields directly.
Outcome enums (RefreshOutcome / RevokeOutcome / RotationOutcome) stay
in crate::error for cohesion with crate::Error (they carry it).

Closes Tech Spec §5.4 file-split commitment.

Tests: 225/225 still pass. Workspace clean. Consumer crates (action/sdk/
engine/plugin/sandbox) compile unchanged.
