+++ crates/resource/src/manager/mod.rs
@@ -1,8 +1,8 @@
 //! Central resource manager — registration, acquire dispatch, and shutdown.
 //!
 //! [`Manager`] is the single entry point for the resource subsystem. It owns
-//! the [`Registry`], [`RecoveryGroupRegistry`], and a [`CancellationToken`]
-//! for coordinated shutdown.
+//! the registry, recovery-group registry, and a [`CancellationToken`] for
+//! coordinated shutdown.
 //!
 //! # Lifecycle
 //!
@@ -13,12 +13,21 @@
 //!   ├── remove()     — unregister + cleanup
 //!   └── shutdown()   — cancel all, drain
 //! ```
+//!
+//! # Submodule layout (Tech Spec §5.4)
+//!
+//! - `options` — `ManagerConfig`, `RegisterOptions`, `ShutdownConfig`, `DrainTimeoutPolicy`
+//! - `registration` — `register_inner` private helper + reverse-index write
+//! - `gate` — `GateAdmission` + `admit_through_gate` + `settle_gate_admission`
+//! - `execute` — resilience pipeline + register-time pool config validation
+//! - `rotation` — `ResourceDispatcher` trampoline + `on_credential_*` fan-out
+//! - `shutdown` — `graceful_shutdown` + drain helpers + `set_phase_all*`
 
 use std::{
     any::TypeId,
     sync::{
         Arc,
-        atomic::{AtomicBool, AtomicU64, Ordering as AtomicOrdering},
+        atomic::{AtomicBool, AtomicU64},
     },
     time::{Duration, Instant},
 };
@@ -30,13 +39,13 @@ use tokio_util::sync::CancellationToken;
 
 use crate::{
     context::ResourceContext,
-    error::{Error, RefreshOutcome, RevokeOutcome},
+    error::Error,
     events::ResourceEvent,
     integration::AcquireResilience,
     metrics::{ResourceOpsMetrics, ResourceOpsSnapshot},
     options::AcquireOptions,
     recovery::{
-        gate::{GateState, RecoveryGate, RecoveryTicket, TryBeginError},
+        gate::{GateState, RecoveryGate},
         group::RecoveryGroupRegistry,
     },
     registry::Registry,
@@ -45,6 +54,18 @@ use crate::{
     resource::Resource,
     runtime::{TopologyRuntime, managed::ManagedResource},
 };
+
+mod execute;
+mod gate;
+pub(crate) mod options;
+mod registration;
+pub(crate) mod rotation;
+mod shutdown;
+
+use execute::{execute_with_resilience, validate_pool_config};
+use gate::{admit_through_gate, settle_gate_admission};
+pub use options::{DrainTimeoutPolicy, ManagerConfig, RegisterOptions, ShutdownConfig};
+pub use shutdown::{ShutdownError, ShutdownReport};
 
 /// Snapshot of a resource's health and operational state.
 #[derive(Debug, Clone)]
@@ -61,242 +82,6 @@ pub struct ResourceHealthSnapshot {
     pub generation: u64,
 }
 
-/// Policy that controls what `graceful_shutdown` does when the
-/// drain phase expires with handles still outstanding (#302).
-///
-/// Before this split, `graceful_shutdown` always proceeded to
-/// `registry.clear()` even on timeout, dropping live `ManagedResource`s
-/// while handles remained outstanding. That turned a cooperative shutdown
-/// into a use-after-logical-drop. The policy makes the choice explicit.
-#[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum DrainTimeoutPolicy {
-    /// On drain timeout, return
-    /// [`ShutdownError::DrainTimeout`] **without** clearing the registry.
-    /// Live handles remain valid and the caller decides what to do next.
-    /// This is the default — it preserves the "graceful" guarantee.
-    #[default]
-    Abort,
-    /// On drain timeout, log, clear the registry anyway, and report the
-    /// outstanding-handle count in [`ShutdownReport`]. Opt-in escape hatch
-    /// for supervisors that must exit on a deadline regardless of cost.
-    Force,
-}
-
-/// Configuration for graceful shutdown.
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub struct ShutdownConfig {
-    /// How long to wait for in-flight handles to be released.
-    pub drain_timeout: Duration,
-    /// What to do on drain timeout. Default: [`DrainTimeoutPolicy::Abort`].
-    pub on_drain_timeout: DrainTimeoutPolicy,
-    /// Upper bound on how long Phase 4 will wait for release-queue
-    /// workers to finish processing outstanding tasks.
-    pub release_queue_timeout: Duration,
-}
-
-impl Default for ShutdownConfig {
-    fn default() -> Self {
-        Self {
-            drain_timeout: Duration::from_secs(30),
-            on_drain_timeout: DrainTimeoutPolicy::Abort,
-            release_queue_timeout: Duration::from_secs(10),
-        }
-    }
-}
-
-impl ShutdownConfig {
-    /// Override the drain timeout, returning `self` for chaining.
-    ///
-    /// `#[non_exhaustive]` prevents external crates from using struct
-    /// literal construction; this (and the sibling setters) is the
-    /// forward-compatible entry point for per-field customization.
-    #[must_use]
-    pub fn with_drain_timeout(mut self, timeout: Duration) -> Self {
-        self.drain_timeout = timeout;
-        self
-    }
-
-    /// Override the drain-timeout policy.
-    #[must_use]
-    pub fn with_drain_timeout_policy(mut self, policy: DrainTimeoutPolicy) -> Self {
-        self.on_drain_timeout = policy;
-        self
-    }
-
-    /// Override the release-queue timeout budget for Phase 4.
-    #[must_use]
-    pub fn with_release_queue_timeout(mut self, timeout: Duration) -> Self {
-        self.release_queue_timeout = timeout;
-        self
-    }
-}
-
-/// Structured result of a successful (or forced-through) graceful shutdown.
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub struct ShutdownReport {
-    /// How many `ResourceGuard`s were still outstanding when the drain
-    /// phase finished. Zero on the happy path. Nonzero only when the
-    /// caller explicitly opted into [`DrainTimeoutPolicy::Force`].
-    pub outstanding_handles_after_drain: u64,
-    /// Whether Phase 3 (`registry.clear`) actually ran.
-    pub registry_cleared: bool,
-    /// Whether Phase 4 (release-queue drain) completed within
-    /// `release_queue_timeout`.
-    pub release_queue_drained: bool,
-}
-
-/// Errors returned by [`Manager::graceful_shutdown`].
-///
-/// Each variant corresponds to a failure mode that was previously silently
-/// absorbed by the old infallible signature. A timeout during drain, for
-/// example, used to be a `tracing::warn!` and a forced `registry.clear()`;
-/// it is now a typed error that the caller must handle.
-#[derive(Debug, thiserror::Error)]
-#[non_exhaustive]
-pub enum ShutdownError {
-    /// `graceful_shutdown` was already in progress when this call entered.
-    /// CAS-guarded so exactly one caller wins the race.
-    #[error("graceful shutdown already in progress")]
-    AlreadyShuttingDown,
-
-    /// The drain phase did not finish within `drain_timeout` and the
-    /// policy was [`DrainTimeoutPolicy::Abort`]. The registry was **not**
-    /// cleared and any outstanding handles remain valid, but every
-    /// registered resource is transitioned to
-    /// [`ResourcePhase::Failed`](crate::state::ResourcePhase::Failed) so
-    /// subsequent acquires fail fast and `health_check` reflects the
-    /// post-abort reality (R-023).
-    #[error(
-        "drain timeout expired with {outstanding} handle(s) still active; registry was NOT cleared (policy=Abort)"
-    )]
-    DrainTimeout {
-        /// Snapshot of the drain-tracker counter at the moment the timeout
-        /// fired.
-        outstanding: u64,
-    },
-
-    /// Phase 4 did not finish within `release_queue_timeout`.
-    #[error("release queue workers did not finish within {timeout:?}")]
-    ReleaseQueueTimeout {
-        /// The budget that was exceeded.
-        timeout: Duration,
-    },
-}
-
-/// Internal drain-phase error used by the private `wait_for_drain` helper.
-/// Carries the outstanding-handle count at the moment the drain timer fired.
-#[derive(Debug)]
-struct DrainTimeoutError {
-    outstanding: u64,
-}
-
-/// Configuration for the [`Manager`].
-#[derive(Debug, Clone)]
-pub struct ManagerConfig {
-    /// Number of background workers for the release queue.
-    ///
-    /// Defaults to 2.
-    pub release_queue_workers: usize,
-    /// Optional shared metrics registry for telemetry counters.
-    ///
-    /// When `Some`, the manager records resource operation counters
-    /// (`acquire_total`, `release_total`, etc.) into the registry.
-    /// When `None`, metrics are silently skipped (zero overhead).
-    pub metrics_registry: Option<Arc<nebula_telemetry::metrics::MetricsRegistry>>,
-    /// Default per-resource timeout budget for credential rotation hooks
-    /// (`on_credential_refresh` / `on_credential_revoke`).
-    ///
-    /// Each registered resource may override this via
-    /// `RegisterOptions::credential_rotation_timeout` (Task 6). When a
-    /// rotation hook exceeds the per-resource budget the dispatcher reports
-    /// `RefreshOutcome::TimedOut` / `RevokeOutcome::TimedOut` and the
-    /// remaining sibling dispatches continue unaffected (security amendment
-    /// B-1: per-resource isolation).
-    ///
-    /// Defaults to 30 seconds.
-    pub credential_rotation_timeout: Duration,
-}
-
-impl Default for ManagerConfig {
-    fn default() -> Self {
-        Self {
-            release_queue_workers: 2,
-            metrics_registry: None,
-            credential_rotation_timeout: Duration::from_secs(30),
-        }
-    }
-}
-
-/// Extended options for resource registration.
-///
-/// Used with the `register_*_with` convenience methods to configure
-/// resilience and recovery beyond the simple `register_*` defaults.
-#[derive(Debug, Clone)]
-pub struct RegisterOptions {
-    /// Scope level for the resource (default: `Global`).
-    pub scope: ScopeLevel,
-    /// Optional acquire resilience (timeout + retry + circuit breaker).
-    pub resilience: Option<AcquireResilience>,
-    /// Optional recovery gate for thundering-herd prevention.
-    pub recovery_gate: Option<Arc<RecoveryGate>>,
-    /// Credential ID this resource binds to.
-    ///
-    /// Required for resources where `R::Credential != NoCredential` —
-    /// `Manager::register` returns [`Error::missing_credential_id`] if a
-    /// credential-bearing resource is registered without an ID. Ignored for
-    /// `NoCredential`-bound resources (the manager logs a warning if one is
-    /// supplied alongside `Credential = NoCredential`).
-    ///
-    /// Set via [`RegisterOptions::with_credential_id`].
-    pub credential_id: Option<CredentialId>,
-    /// Per-resource override for the default credential rotation timeout.
-    ///
-    /// `None` falls back to [`ManagerConfig::credential_rotation_timeout`]
-    /// (default `30s`). Only meaningful for credential-bearing resources;
-    /// ignored for `NoCredential`-bound resources.
-    ///
-    /// Set via [`RegisterOptions::with_rotation_timeout`].
-    pub credential_rotation_timeout: Option<Duration>,
-}
-
-impl Default for RegisterOptions {
-    fn default() -> Self {
-        Self {
-            scope: ScopeLevel::Global,
-            resilience: None,
-            recovery_gate: None,
-            credential_id: None,
-            credential_rotation_timeout: None,
-        }
-    }
-}
-
-impl RegisterOptions {
-    /// Sets the credential ID this resource binds to.
-    ///
-    /// Required for credential-bearing resources (`R::Credential != NoCredential`).
-    /// `Manager::register` errors with [`Error::missing_credential_id`] if a
-    /// credential-bearing resource is registered without an ID.
-    #[must_use]
-    pub fn with_credential_id(mut self, id: CredentialId) -> Self {
-        self.credential_id = Some(id);
-        self
-    }
-
-    /// Overrides the default credential rotation timeout for this resource.
-    ///
-    /// Falls back to [`ManagerConfig::credential_rotation_timeout`] (default
-    /// `30s`) when not set. Only meaningful for credential-bearing resources.
-    #[must_use]
-    pub fn with_rotation_timeout(mut self, timeout: Duration) -> Self {
-        self.credential_rotation_timeout = Some(timeout);
-        self
-    }
-}
-
 /// Central registry and lifecycle manager for all resources.
 ///
 /// Owns the [`ReleaseQueue`] internally — callers never need to create,
@@ -306,19 +91,19 @@ impl RegisterOptions {
 /// Thread-safe: all internal state is behind concurrent data structures.
 /// Share via `Arc<Manager>` across tasks.
 pub struct Manager {
-    registry: Registry,
-    recovery_groups: RecoveryGroupRegistry,
-    cancel: CancellationToken,
-    metrics: Option<ResourceOpsMetrics>,
-    event_tx: broadcast::Sender<ResourceEvent>,
-    release_queue: Arc<ReleaseQueue>,
-    release_queue_handle: tokio::sync::Mutex<Option<ReleaseQueueHandle>>,
+    pub(super) registry: Registry,
+    pub(super) recovery_groups: RecoveryGroupRegistry,
+    pub(super) cancel: CancellationToken,
+    pub(super) metrics: Option<ResourceOpsMetrics>,
+    pub(super) event_tx: broadcast::Sender<ResourceEvent>,
+    pub(super) release_queue: Arc<ReleaseQueue>,
+    pub(super) release_queue_handle: tokio::sync::Mutex<Option<ReleaseQueueHandle>>,
     /// Tracks active `ResourceGuard`s for drain-aware shutdown.
-    drain_tracker: Arc<(AtomicU64, Notify)>,
+    pub(super) drain_tracker: Arc<(AtomicU64, Notify)>,
     /// CAS-guarded idempotency flag for `graceful_shutdown`. Flipped
     /// false → true by the winning caller; losers return
     /// [`ShutdownError::AlreadyShuttingDown`].
-    shutting_down: AtomicBool,
+    pub(super) shutting_down: AtomicBool,
     /// Reverse index: credential_id → dispatchers for resources that bind to this credential.
     ///
     /// Populated at register time when `R::Credential != NoCredential`. Read by
@@ -326,17 +111,17 @@ pub struct Manager {
     /// in parallel via `join_all` (per Tech Spec §3.2).
     ///
     /// `Arc<dyn ResourceDispatcher>` provides type-erased dispatch — see
-    /// `crate::rotation::TypedDispatcher<R>`.
-    credential_resources:
-        dashmap::DashMap<CredentialId, Vec<Arc<dyn crate::rotation::ResourceDispatcher>>>,
+    /// `crate::manager::rotation::TypedDispatcher<R>`.
+    pub(super) credential_resources:
+        dashmap::DashMap<CredentialId, Vec<Arc<dyn rotation::ResourceDispatcher>>>,
     /// Default per-resource timeout budget for credential rotation hooks.
     ///
     /// Sourced from [`ManagerConfig::credential_rotation_timeout`] at
     /// construction. Per-resource overrides flow through
     /// `TypedDispatcher::timeout_override()` (set at register time, Task 6).
-    credential_rotation_timeout: Duration,
+    pub(super) credential_rotation_timeout: Duration,
     /// Optional lifecycle handle for coordinated cancellation (spec 08).
-    lifecycle: Option<LayerLifecycle>,
+    pub(super) lifecycle: Option<LayerLifecycle>,
 }
 
 impl Manager {
@@ -433,14 +218,9 @@ impl Manager {
     /// Returns an error if config validation fails on the provided config,
     /// or if a credential-bearing resource is registered without a
     /// `credential_id` (see [`Error::missing_credential_id`]).
-    // Task 11 (Manager file-split) is slated to consolidate scope/resilience/
-    // recovery_gate/credential_id/credential_rotation_timeout into a single
-    // `RegisterOptions` parameter. Until that lands, the positional surface
-    // mirrors the existing pre-Task-6 signature with two new credential
-    // parameters wired through to `register_inner`.
     #[expect(
         clippy::too_many_arguments,
-        reason = "consolidation into `RegisterOptions` deferred to Task 11 (Manager file-split); self-fires when arg count drops"
+        reason = "consolidation into `RegisterOptions` deferred; the `_with` shorthand variants already use the consolidated form"
     )]
     pub fn register<R: Resource>(
         &self,
@@ -498,52 +278,6 @@ impl Manager {
         )?;
 
         tracing::debug!(%key, "resource registered");
-        Ok(())
-    }
-
-    /// Internal helper: write to `credential_resources` reverse-index when the
-    /// resource binds a real credential; no-op for `NoCredential`-bound resources.
-    ///
-    /// Called by `register<R>` after the registry write succeeds. The `TypeId`
-    /// check distinguishes credential-bearing resources from opt-out marker types
-    /// at compile time.
-    ///
-    /// # Errors
-    ///
-    /// Returns `Error::missing_credential_id` when a credential-bearing resource
-    /// (`R::Credential != NoCredential`) is registered without a `credential_id`.
-    fn register_inner<R: Resource>(
-        &self,
-        managed: Arc<ManagedResource<R>>,
-        credential_id: Option<CredentialId>,
-        timeout_override: Option<Duration>,
-    ) -> Result<(), Error> {
-        let opted_out =
-            TypeId::of::<R::Credential>() == TypeId::of::<nebula_credential::NoCredential>();
-
-        match (opted_out, credential_id) {
-            (true, Some(_)) => {
-                tracing::warn!(
-                    resource = %R::key(),
-                    "register: NoCredential resource provided a credential_id; ignoring"
-                );
-            },
-            (true, None) => {
-                // Normal path for NoCredential-bound resources — no reverse-index write.
-            },
-            (false, None) => {
-                return Err(Error::missing_credential_id(R::key()));
-            },
-            (false, Some(id)) => {
-                let dispatcher: Arc<dyn crate::rotation::ResourceDispatcher> = Arc::new(
-                    crate::rotation::TypedDispatcher::new(managed, timeout_override),
-                );
-                self.credential_resources
-                    .entry(id)
-                    .or_default()
-                    .push(dispatcher);
-            },
-        }
         Ok(())
     }
 
@@ -1051,11 +785,6 @@ impl Manager {
         })
         .await;
 
-        // Settle the gate ticket based on the acquire result. #322: this
-        // makes the ticket ownership end-to-end — on success we `resolve`,
-        // on retryable error we `fail_transient`, on permanent error we
-        // `fail_permanent`. The `Drop` impl of `RecoveryTicket` covers
-        // cancellation/panic paths.
         settle_gate_admission(gate_admission, &result);
         self.record_acquire_result(&result, started);
         result.map(|h| h.with_drain_tracker(self.drain_tracker.clone()))
@@ -1131,11 +860,6 @@ impl Manager {
         })
         .await;
 
-        // Settle the gate ticket based on the acquire result. #322: this
-        // makes the ticket ownership end-to-end — on success we `resolve`,
-        // on retryable error we `fail_transient`, on permanent error we
-        // `fail_permanent`. The `Drop` impl of `RecoveryTicket` covers
-        // cancellation/panic paths.
         settle_gate_admission(gate_admission, &result);
         self.record_acquire_result(&result, started);
         result.map(|h| h.with_drain_tracker(self.drain_tracker.clone()))
@@ -1212,11 +936,6 @@ impl Manager {
         })
         .await;
 
-        // Settle the gate ticket based on the acquire result. #322: this
-        // makes the ticket ownership end-to-end — on success we `resolve`,
-        // on retryable error we `fail_transient`, on permanent error we
-        // `fail_permanent`. The `Drop` impl of `RecoveryTicket` covers
-        // cancellation/panic paths.
         settle_gate_admission(gate_admission, &result);
         self.record_acquire_result(&result, started);
         result.map(|h| h.with_drain_tracker(self.drain_tracker.clone()))
@@ -1292,11 +1011,6 @@ impl Manager {
         })
         .await;
 
-        // Settle the gate ticket based on the acquire result. #322: this
-        // makes the ticket ownership end-to-end — on success we `resolve`,
-        // on retryable error we `fail_transient`, on permanent error we
-        // `fail_permanent`. The `Drop` impl of `RecoveryTicket` covers
-        // cancellation/panic paths.
         settle_gate_admission(gate_admission, &result);
         self.record_acquire_result(&result, started);
         result.map(|h| h.with_drain_tracker(self.drain_tracker.clone()))
@@ -1377,11 +1091,6 @@ impl Manager {
             ))),
         };
 
-        // Settle the gate ticket based on the acquire result. #322: this
-        // makes the ticket ownership end-to-end — on success we `resolve`,
-        // on retryable error we `fail_transient`, on permanent error we
-        // `fail_permanent`. The `Drop` impl of `RecoveryTicket` covers
-        // cancellation/panic paths.
         settle_gate_admission(gate_admission, &result);
         self.record_acquire_result(&result, started);
         result.map(|h| h.with_drain_tracker(self.drain_tracker.clone()))
@@ -1620,273 +1329,6 @@ impl Manager {
         Ok(outcome)
     }
 
-    /// Handle a credential refresh event by fanning out
-    /// `Resource::on_credential_refresh` hooks to every resource bound to
-    /// `credential_id`.
-    ///
-    /// Each per-resource future runs concurrently via `join_all` and has its
-    /// own timeout budget — one slow or failing resource never poisons
-    /// siblings (security amendment B-1 isolation). The per-resource budget
-    /// is sourced from `RegisterOptions::credential_rotation_timeout` if set
-    /// at register time, otherwise the manager's
-    /// [`ManagerConfig::credential_rotation_timeout`] default.
-    ///
-    /// Returns a list of `(ResourceKey, RefreshOutcome)` pairs — one per
-    /// affected resource. If no resources are bound to the credential the
-    /// result is an empty `Vec`.
-    ///
-    /// # Caller contract
-    ///
-    /// The caller (typically the engine refresh coordinator) constructs a
-    /// [`SchemeFactory<C>`](nebula_credential::SchemeFactory) over the
-    /// freshly-projected post-refresh state. Per fan-out branch the manager
-    /// clones the factory, type-erases it through a `Box<dyn Any>`, and
-    /// passes it to the resource's typed dispatcher. The dispatcher
-    /// downcasts back to `SchemeFactory<R::Credential>` and calls
-    /// `factory.acquire().await` to mint a fresh
-    /// [`SchemeGuard<'_, R::Credential>`](nebula_credential::SchemeGuard)
-    /// inside its own typed scope before invoking the resource hook.
-    ///
-    /// **Why dispatcher-side acquire (not manager-side):** `Box<dyn Any>`
-    /// is `'static`-bound (because `Any: 'static`), so a non-`'static`
-    /// `SchemeGuard<'_, C>` cannot be type-erased through it. `SchemeFactory<C>`
-    /// IS `'static`, so it can. The acquire call is the same in either
-    /// design (one per dispatcher); only the call boundary moves.
-    ///
-    /// # Errors
-    ///
-    /// This call itself never returns `Err`; per-resource failures are
-    /// reported in the returned `Vec` via [`RefreshOutcome::Failed`] and
-    /// [`RefreshOutcome::TimedOut`]. The `Result` shape is preserved for
-    /// forward-compat with future caller-level guards (e.g. shutdown-in-
-    /// progress short-circuit).
-    pub async fn on_credential_refreshed<C: Credential>(
-        &self,
-        credential_id: &CredentialId,
-        factory: nebula_credential::SchemeFactory<C>,
-        ctx: &nebula_credential::CredentialContext,
-    ) -> Result<Vec<(ResourceKey, RefreshOutcome)>, Error> {
-        use futures::future::join_all;
-
-        let dispatchers = self
-            .credential_resources
-            .get(credential_id)
-            .map(|entry| entry.value().clone())
-            .unwrap_or_default();
-
-        if dispatchers.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        use tracing::Instrument as _;
-
-        let scheme_type_id = TypeId::of::<<C as Credential>::Scheme>();
-        let default_timeout = self.credential_rotation_timeout;
-        let metrics = self.metrics.clone();
-
-        let span = tracing::info_span!(
-            "resource.credential_refresh",
-            credential_id = %credential_id,
-            resources_affected = dispatchers.len(),
-        );
-
-        // Per-resource futures with isolation. Each future has its own
-        // timeout budget; one slow or failing resource never poisons
-        // siblings (security amendment B-1).
-        let futures = dispatchers.into_iter().filter_map(|d| {
-            // Defensive scheme-type check. The dispatcher's
-            // `scheme_type_id()` should match the call's `C` if
-            // `register_inner` was correct; if not, log and skip rather
-            // than panic.
-            if d.scheme_type_id() != scheme_type_id {
-                tracing::error!(
-                    resource = %d.resource_key(),
-                    expected_scheme_type = ?scheme_type_id,
-                    got_scheme_type = ?d.scheme_type_id(),
-                    "dispatcher scheme_type_id mismatch — skipping (register_inner bug?)",
-                );
-                return None;
-            }
-
-            let timeout = d.timeout_override().unwrap_or(default_timeout);
-            let key = d.resource_key();
-            // Cheap Arc bump per dispatcher; the boxed factory becomes the
-            // dispatcher's owned `SchemeFactory<R::Credential>` after
-            // downcast.
-            let factory_box: Box<dyn std::any::Any + Send + Sync> = Box::new(factory.clone());
-            let metrics = metrics.clone();
-
-            Some(async move {
-                let dispatch_started = Instant::now();
-                let dispatch = d.dispatch_refresh(factory_box, ctx);
-                let outcome = match tokio::time::timeout(timeout, dispatch).await {
-                    Ok(Ok(())) => RefreshOutcome::Ok,
-                    Ok(Err(e)) => RefreshOutcome::Failed(e),
-                    Err(_) => RefreshOutcome::TimedOut { budget: timeout },
-                };
-                let elapsed = dispatch_started.elapsed().as_secs_f64();
-
-                if let Some(m) = metrics.as_ref() {
-                    m.record_rotation_dispatch(&outcome, elapsed);
-                }
-
-                (key, outcome)
-            })
-        });
-
-        let results: Vec<(ResourceKey, RefreshOutcome)> = join_all(futures).instrument(span).await;
-
-        // Aggregate per-resource outcomes into a `RotationOutcome` for the
-        // `CredentialRefreshed` event payload (Tech Spec §6.2). Per-resource
-        // dispatch metrics are emitted inline above; this is the cycle-level
-        // aggregate.
-        let mut outcome = crate::error::RotationOutcome::default();
-        for (_, o) in &results {
-            match o {
-                RefreshOutcome::Ok => outcome.ok += 1,
-                RefreshOutcome::Failed(_) => outcome.failed += 1,
-                RefreshOutcome::TimedOut { .. } => outcome.timed_out += 1,
-            }
-        }
-        // Broadcast send errors (no live subscribers) are intentionally ignored.
-        let _ = self.event_tx.send(ResourceEvent::CredentialRefreshed {
-            credential_id: *credential_id,
-            resources_affected: results.len(),
-            outcome,
-        });
-
-        Ok(results)
-    }
-
-    /// Handle a credential revocation event by fanning out
-    /// `Resource::on_credential_revoke` hooks to every resource bound to
-    /// `credential_id`.
-    ///
-    /// Symmetric to [`on_credential_refreshed`](Self::on_credential_refreshed)
-    /// minus the scheme-material plumbing — `on_credential_revoke` takes only
-    /// `&CredentialId`, so no `SchemeFactory<C>` or generic `C: Credential`
-    /// bound is required here.
-    ///
-    /// Each per-resource future runs concurrently via `join_all` and has its
-    /// own timeout budget — one slow or failing resource never poisons
-    /// siblings (security amendment B-1 isolation). The per-resource budget
-    /// is sourced from `RegisterOptions::credential_rotation_timeout` if set
-    /// at register time, otherwise the manager's
-    /// [`ManagerConfig::credential_rotation_timeout`] default.
-    ///
-    /// Per security amendment B-2, every non-`Ok` per-resource outcome
-    /// (`Failed` or `TimedOut`) emits a
-    /// [`ResourceEvent::HealthChanged`] event with `healthy: false` inline
-    /// so the operator sees a per-resource failure signal even if the
-    /// aggregate `CredentialRevoked` event (Task 7) is dropped by a
-    /// saturated subscriber. Successful revocations emit only the aggregate
-    /// event.
-    ///
-    /// Returns a list of `(ResourceKey, RevokeOutcome)` pairs — one per
-    /// affected resource. If no resources are bound to the credential the
-    /// result is an empty `Vec`.
-    ///
-    /// # Errors
-    ///
-    /// This call itself never returns `Err`; per-resource failures are
-    /// reported in the returned `Vec` via [`RevokeOutcome::Failed`] and
-    /// [`RevokeOutcome::TimedOut`]. The `Result` shape is preserved for
-    /// forward-compat with future caller-level guards (e.g. shutdown-in-
-    /// progress short-circuit).
-    pub async fn on_credential_revoked(
-        &self,
-        credential_id: &CredentialId,
-    ) -> Result<Vec<(ResourceKey, RevokeOutcome)>, Error> {
-        use futures::future::join_all;
-
-        let dispatchers = self
-            .credential_resources
-            .get(credential_id)
-            .map(|entry| entry.value().clone())
-            .unwrap_or_default();
-
-        if dispatchers.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        use tracing::Instrument as _;
-
-        let default_timeout = self.credential_rotation_timeout;
-        let event_tx = self.event_tx.clone();
-        let metrics = self.metrics.clone();
-
-        let span = tracing::warn_span!(
-            "resource.credential_revoke",
-            credential_id = %credential_id,
-            resources_affected = dispatchers.len(),
-        );
-
-        // Per-resource futures with isolation. Each future has its own
-        // timeout budget; one slow or failing resource never poisons
-        // siblings (security amendment B-1). Per security amendment B-2:
-        // emit HealthChanged{healthy:false} inline for any non-Ok outcome
-        // so the operator sees a per-resource failure signal even if the
-        // aggregate event is lost.
-        let futures = dispatchers.into_iter().map(|d| {
-            let timeout = d.timeout_override().unwrap_or(default_timeout);
-            let key = d.resource_key();
-            let credential_id = *credential_id;
-            let event_tx = event_tx.clone();
-            let metrics = metrics.clone();
-
-            async move {
-                let dispatch_started = Instant::now();
-                let dispatch = d.dispatch_revoke(&credential_id);
-                let outcome = match tokio::time::timeout(timeout, dispatch).await {
-                    Ok(Ok(())) => RevokeOutcome::Ok,
-                    Ok(Err(e)) => RevokeOutcome::Failed(e),
-                    Err(_) => RevokeOutcome::TimedOut { budget: timeout },
-                };
-                let elapsed = dispatch_started.elapsed().as_secs_f64();
-
-                if let Some(m) = metrics.as_ref() {
-                    m.record_revoke_dispatch(&outcome, elapsed);
-                }
-
-                // Security amendment B-2: emit HealthChanged{healthy:false}
-                // for non-Ok outcomes. Successful revocations emit only the
-                // aggregate CredentialRevoked event below. Broadcast send
-                // errors (no live subscribers) are intentionally ignored.
-                if !matches!(outcome, RevokeOutcome::Ok) {
-                    let _ = event_tx.send(ResourceEvent::HealthChanged {
-                        key: key.clone(),
-                        healthy: false,
-                    });
-                }
-
-                (key, outcome)
-            }
-        });
-
-        let results: Vec<(ResourceKey, RevokeOutcome)> = join_all(futures).instrument(span).await;
-
-        // Aggregate per-resource outcomes into a `RotationOutcome` for the
-        // `CredentialRevoked` event payload (Tech Spec §6.2). Per-resource
-        // dispatch metrics are emitted inline above; this is the cycle-level
-        // aggregate.
-        let mut outcome = crate::error::RotationOutcome::default();
-        for (_, o) in &results {
-            match o {
-                RevokeOutcome::Ok => outcome.ok += 1,
-                RevokeOutcome::Failed(_) => outcome.failed += 1,
-                RevokeOutcome::TimedOut { .. } => outcome.timed_out += 1,
-            }
-        }
-        // Broadcast send errors (no live subscribers) are intentionally ignored.
-        let _ = self.event_tx.send(ResourceEvent::CredentialRevoked {
-            credential_id: *credential_id,
-            resources_affected: results.len(),
-            outcome,
-        });
-
-        Ok(results)
-    }
-
     /// Removes a resource from the registry by key.
     ///
     /// # Errors
@@ -1917,233 +1359,6 @@ impl Manager {
     pub fn shutdown(&self) {
         tracing::info!("resource manager shutting down");
         self.cancel.cancel();
-    }
-
-    /// Triggers graceful shutdown with drain and cleanup.
-    ///
-    /// 1. **Signal** — cancels the token so new acquires are rejected.
-    /// 2. **Drain** — waits up to [`ShutdownConfig::drain_timeout`] for in-flight handles to be
-    ///    released.
-    /// 3. **Clear** — drops all managed resources, releasing their `Arc<ReleaseQueue>` references
-    ///    so workers can drain and exit.
-    /// 4. **Await workers** — waits for the release queue workers to finish processing remaining
-    ///    tasks.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// # use nebula_resource::manager::{Manager, ShutdownConfig};
-    /// # use std::time::Duration;
-    /// # async fn example() {
-    /// let manager = Manager::new();
-    /// manager
-    ///     .graceful_shutdown(ShutdownConfig::default().with_drain_timeout(Duration::from_secs(5)))
-    ///     .await
-    ///     .expect("graceful shutdown should succeed");
-    /// # }
-    /// ```
-    pub async fn graceful_shutdown(
-        &self,
-        config: ShutdownConfig,
-    ) -> Result<ShutdownReport, ShutdownError> {
-        // CAS idempotency guard: exactly one caller wins. Concurrent callers
-        // that arrive after this CAS see `AlreadyShuttingDown` immediately
-        // rather than re-entering the drain logic against a half-torn state.
-        if self
-            .shutting_down
-            .compare_exchange(false, true, AtomicOrdering::AcqRel, AtomicOrdering::Acquire)
-            .is_err()
-        {
-            return Err(ShutdownError::AlreadyShuttingDown);
-        }
-
-        tracing::info!("resource manager: starting graceful shutdown");
-
-        // Phase 1: SIGNAL — cancel the shared token. This does two things:
-        //   a) Rejects new acquire calls (checked in `lookup`).
-        //   b) Tells release queue workers to drain remaining tasks and exit
-        //      (they share this token via `ReleaseQueue::with_cancel`).
-        self.cancel.cancel();
-
-        // #387: mark every registered resource as `Draining` so operators
-        // polling `health_check` during the drain window see the correct
-        // lifecycle phase instead of a stale `Ready`.
-        self.set_phase_all(crate::state::ResourcePhase::Draining);
-
-        // Phase 2: DRAIN — wait for in-flight handles to be released.
-        // On timeout, respect the policy: Abort preserves "graceful"
-        // (returns Err *without* clearing the registry), Force proceeds
-        // but records the outstanding count in the report.
-        let mut outstanding_after_drain: u64 = 0;
-        match self.wait_for_drain(config.drain_timeout).await {
-            Ok(()) => {},
-            Err(DrainTimeoutError { outstanding }) => match config.on_drain_timeout {
-                DrainTimeoutPolicy::Abort => {
-                    tracing::warn!(
-                        outstanding,
-                        "resource manager: drain timeout, policy=Abort — \
-                         registry preserved, marking all resources Failed, \
-                         returning DrainTimeout"
-                    );
-                    // R-023 / 🔴-4: every resource transitions to `Failed`
-                    // (with `HealthChanged{healthy:false}` emitted per key)
-                    // rather than back to `Ready`. The cancel token fired
-                    // in Phase 1 already rejects new acquires; pretending
-                    // the registry is `Ready` while the caller observes a
-                    // `DrainTimeout` is phase corruption — callers polling
-                    // `health_check` would see `Ready` but get
-                    // `Error::cancelled` from `lookup`. Marking `Failed`
-                    // makes the registry tell the truth.
-                    let err = ShutdownError::DrainTimeout { outstanding };
-                    self.set_phase_all_failed(&err);
-                    self.shutting_down.store(false, AtomicOrdering::Release);
-                    return Err(err);
-                },
-                DrainTimeoutPolicy::Force => {
-                    tracing::warn!(
-                        outstanding,
-                        "resource manager: drain timeout, policy=Force — \
-                         clearing registry anyway"
-                    );
-                    outstanding_after_drain = outstanding;
-                },
-            },
-        }
-
-        // #387: drain has completed (or been force-released). Mark every
-        // resource as `ShuttingDown` so a health snapshot captured in the
-        // narrow window between here and `registry.clear()` reflects the
-        // real lifecycle state.
-        self.set_phase_all(crate::state::ResourcePhase::ShuttingDown);
-
-        // Phase 3: CLEAR — drop all ManagedResources so their
-        // Arc<ReleaseQueue> refs are released.
-        self.registry.clear();
-
-        // Phase 4: AWAIT WORKERS — workers are already draining (from
-        // Phase 1 cancel signal). Await with a bounded timeout; failure
-        // to finish in time is a typed error, not a swallowed warning.
-        if let Some(handle) = self.release_queue_handle.lock().await.take() {
-            let shutdown_fut = ReleaseQueue::shutdown(handle);
-            if tokio::time::timeout(config.release_queue_timeout, shutdown_fut)
-                .await
-                .is_err()
-            {
-                tracing::warn!(
-                    timeout = ?config.release_queue_timeout,
-                    "resource manager: release queue workers did not \
-                     finish within release_queue_timeout"
-                );
-                return Err(ShutdownError::ReleaseQueueTimeout {
-                    timeout: config.release_queue_timeout,
-                });
-            }
-        }
-
-        tracing::info!("resource manager: shutdown complete");
-        Ok(ShutdownReport {
-            outstanding_handles_after_drain: outstanding_after_drain,
-            registry_cleared: true,
-            // If we reached this line Phase 4 either succeeded or had no
-            // work to drain — either way the contract is "drained".
-            release_queue_drained: true,
-        })
-    }
-
-    /// Drives every registered resource to the given lifecycle phase.
-    ///
-    /// Type-erased bulk update used during graceful shutdown so that
-    /// `health_check` returns the correct phase while the drain/cleanup
-    /// is in flight (#387).
-    fn set_phase_all(&self, phase: crate::state::ResourcePhase) {
-        for managed in self.registry.all_managed() {
-            managed.set_phase_erased(phase);
-        }
-    }
-
-    /// Marks every registered resource as `Failed` with the supplied
-    /// shutdown error and emits a per-resource
-    /// [`ResourceEvent::HealthChanged`] with `healthy: false`.
-    ///
-    /// Used by the [`DrainTimeoutPolicy::Abort`] branch (R-023) so the
-    /// registry's recorded phase agrees with the `Err(DrainTimeout)`
-    /// observed by the caller. Without this, `health_check` would report
-    /// `Ready` while `lookup` rejected acquires via the cancel token —
-    /// the exact "phase corruption" called out in Phase 1 finding 🔴-4.
-    ///
-    /// Broadcast send errors (no live subscribers) are intentionally
-    /// ignored, matching the rest of the manager's event-emission policy.
-    fn set_phase_all_failed(&self, error: &ShutdownError) {
-        let reason = error.to_string();
-        for managed in self.registry.all_managed() {
-            managed.set_failed_erased(&reason);
-            let _ = self.event_tx.send(ResourceEvent::HealthChanged {
-                key: managed.resource_key(),
-                healthy: false,
-            });
-        }
-    }
-
-    /// Waits until all active `ResourceGuard`s are dropped or timeout expires.
-    ///
-    /// The loop uses a `register-then-check` ordering to avoid the classic
-    /// `Notify::notify_waiters` lost-wakeup:
-    ///
-    /// 1. Construct + pin + `enable()` a fresh `Notified` future. Calling `enable()` registers this
-    ///    waiter on the `Notify` queue without requiring a `.await`, so any subsequent
-    ///    `notify_waiters()` (fired when a handle's `Drop` decrements the counter from 1 → 0) will
-    ///    reach us.
-    /// 2. Re-check the counter. If it already hit 0 between the outer initial check and our
-    ///    registration, return now — the wakeup we would otherwise wait for has already been
-    ///    consumed.
-    /// 3. Only then await the `Notified` future.
-    ///
-    /// Without this ordering, a burst of handle drops that completes the
-    /// drain *before* the first `notified().await` poll would leak the
-    /// notification entirely, stalling `graceful_shutdown` for the full
-    /// `drain_timeout` (default 30 s) and risking `SIGKILL` escalation
-    /// under a tight orchestrator shutdown window.
-    async fn wait_for_drain(&self, timeout: Duration) -> Result<(), DrainTimeoutError> {
-        let active = self.drain_tracker.0.load(AtomicOrdering::Acquire);
-        if active == 0 {
-            return Ok(());
-        }
-
-        tracing::debug!(active_handles = active, "waiting for handles to drain");
-        let tracker = &self.drain_tracker;
-        let drained = tokio::time::timeout(timeout, async {
-            loop {
-                // Pre-register this waiter BEFORE re-checking the counter.
-                let notified = tracker.1.notified();
-                tokio::pin!(notified);
-                notified.as_mut().enable();
-
-                // Re-check after registration. If the last handle dropped
-                // while we were between the outer check and `enable()`,
-                // the counter is now 0 and we would otherwise wait on a
-                // notification that has already fired.
-                if tracker.0.load(AtomicOrdering::Acquire) == 0 {
-                    return;
-                }
-
-                notified.await;
-
-                if tracker.0.load(AtomicOrdering::Acquire) == 0 {
-                    return;
-                }
-            }
-        })
-        .await;
-
-        if drained.is_err() {
-            let outstanding = tracker.0.load(AtomicOrdering::Acquire);
-            tracing::warn!(
-                outstanding,
-                "resource manager: drain timeout expired with handles still active"
-            );
-            return Err(DrainTimeoutError { outstanding });
-        }
-        Ok(())
     }
 
     /// Returns `true` if a resource with the given key is registered.
@@ -2241,148 +1456,6 @@ impl Manager {
     }
 }
 
-/// Outcome of the pre-acquire gate admission check (#322).
-///
-/// The old `check_recovery_gate` → `trigger_recovery_on_failure` split was
-/// a stampede hazard: after the backoff expired, every caller's `state()`
-/// read returned the same snapshot and all of them proceeded through to
-/// hit the dead backend. The gate was only advanced to `InProgress` *after*
-/// failure, in `trigger_recovery_on_failure`, which is too late.
-///
-/// This enum pushes the CAS-based single-probe claim into the pre-acquire
-/// check: either the caller is granted a ticket (`Probe`) that must be
-/// resolved end-to-end with the acquire result, or the gate was idle/absent
-/// (`Open`), or we got a typed error out directly.
-enum GateAdmission {
-    /// No gate attached. Proceed normally; no ticket ownership.
-    Open,
-    /// Gate attached and currently healthy (`Idle`). Proceed without a
-    /// ticket, but retain the gate so a retryable acquire error can mark it
-    /// failed and open the backoff window for subsequent callers.
-    OpenGated(Arc<RecoveryGate>),
-    /// This caller has been granted the single recovery slot. The acquire
-    /// **must** consume the ticket by calling `resolve()`, `fail_transient`,
-    /// or `fail_permanent` based on the acquire result. Dropping it without
-    /// resolution auto-fails to `GateState::Failed` via its `Drop` impl —
-    /// so even a cancellation or panic in the acquire path is safe.
-    Probe(RecoveryTicket),
-}
-
-/// Admits a caller through the optional recovery gate.
-///
-/// Healthy gates (`Idle`) admit immediately with no CAS so regular traffic
-/// keeps full pool concurrency. Only callers entering while the gate is in a
-/// retryable `Failed` state claim a probe ticket.
-fn admit_through_gate(gate: &Option<Arc<RecoveryGate>>) -> Result<GateAdmission, Error> {
-    let Some(gate) = gate else {
-        return Ok(GateAdmission::Open);
-    };
-
-    match gate.state() {
-        GateState::Idle => Ok(GateAdmission::OpenGated(Arc::clone(gate))),
-        GateState::InProgress { .. } => Err(Error::transient(
-            "backend recovery in progress, retry later",
-        )),
-        GateState::Failed { retry_at, .. } => {
-            if Instant::now() < retry_at {
-                let wait = retry_at.saturating_duration_since(Instant::now());
-                return Err(Error::exhausted("backend recovering", Some(wait)));
-            }
-            match gate.try_begin() {
-                Ok(ticket) => Ok(GateAdmission::Probe(ticket)),
-                Err(TryBeginError::AlreadyInProgress(_waiter)) => Err(Error::transient(
-                    "backend recovery in progress, retry later",
-                )),
-                Err(TryBeginError::RetryLater { retry_at }) => {
-                    let wait = retry_at.saturating_duration_since(Instant::now());
-                    Err(Error::exhausted("backend recovering", Some(wait)))
-                },
-                Err(TryBeginError::PermanentlyFailed { message }) => Err(Error::permanent(message)),
-            }
-        },
-        GateState::PermanentlyFailed { message } => Err(Error::permanent(message)),
-    }
-}
-
-/// Resolves the ticket granted by [`admit_through_gate`] based on the
-/// acquire result. No-op when the admission was [`GateAdmission::Open`]
-/// (no gate attached), so callers can always call this unconditionally.
-fn settle_gate_admission<T>(admission: GateAdmission, result: &Result<T, Error>) {
-    match (admission, result) {
-        (GateAdmission::Probe(ticket), Ok(_)) => ticket.resolve(),
-        (GateAdmission::Probe(ticket), Err(e)) if e.is_retryable() => {
-            ticket.fail_transient(e.to_string());
-        },
-        (GateAdmission::Probe(ticket), Err(_e)) => {
-            // Non-retryable errors are not backend-health signals; keep the
-            // gate open to avoid permanently bricking acquires.
-            ticket.resolve();
-        },
-        (GateAdmission::OpenGated(gate), Err(e)) if e.is_retryable() => {
-            // First retryable failure on healthy path opens the backoff gate.
-            if let Ok(ticket) = gate.try_begin() {
-                ticket.fail_transient(e.to_string());
-            }
-        },
-        (GateAdmission::OpenGated(_) | GateAdmission::Open, _) => {},
-    }
-}
-
-/// Executes an async operation with optional timeout and retry from
-/// [`AcquireResilience`] configuration.
-///
-/// Delegates to [`nebula_resilience::retry_with`] which handles exponential
-/// backoff, wall-clock budget, retry-after hints, and `Classify`-based
-/// error filtering automatically.
-async fn execute_with_resilience<F, Fut, T>(
-    resilience: &Option<AcquireResilience>,
-    mut operation: F,
-) -> Result<T, Error>
-where
-    F: FnMut() -> Fut,
-    Fut: Future<Output = Result<T, Error>> + Send,
-{
-    let Some(config) = resilience else {
-        return operation().await;
-    };
-
-    // #383: `to_retry_config` returns `None` only if the underlying
-    // `RetryConfig::new` rejects the clamped attempt count. That path is
-    // unreachable today; if it ever fires we prefer to fall through to a
-    // single un-retried attempt rather than panic the manager.
-    let Some(retry_cfg) = config.to_retry_config() else {
-        return operation().await;
-    };
-    nebula_resilience::retry_with(retry_cfg, operation)
-        .await
-        .map_err(|call_err| match call_err {
-            nebula_resilience::CallError::Operation(e)
-            | nebula_resilience::CallError::RetriesExhausted { last: e, .. } => e,
-            nebula_resilience::CallError::Timeout(d) => {
-                Error::transient(format!("acquire timed out after {d:?}"))
-            },
-            other => Error::transient(other.to_string()),
-        })
-}
-
-/// Validates pool config invariants at registration time.
-///
-/// Catches obviously broken configs (`max_size == 0`, `min_size > max_size`)
-/// before they reach [`PoolRuntime`], so warmup never inflates beyond
-/// `max_size` and callers cannot deadlock on an empty semaphore (#390).
-fn validate_pool_config(cfg: &crate::topology::pooled::config::Config) -> Result<(), Error> {
-    if cfg.max_size == 0 {
-        return Err(Error::permanent("pool max_size must be > 0"));
-    }
-    if cfg.min_size > cfg.max_size {
-        return Err(Error::permanent(format!(
-            "pool min_size ({}) must be <= max_size ({})",
-            cfg.min_size, cfg.max_size,
-        )));
-    }
-    Ok(())
-}
-
 impl Default for Manager {
     fn default() -> Self {
         Self::new()
@@ -2395,221 +1468,5 @@ impl std::fmt::Debug for Manager {
             .field("registered_count", &self.registry.keys().len())
             .field("is_shutdown", &self.is_shutdown())
             .finish_non_exhaustive()
-    }
-}
-
-#[cfg(test)]
-mod gate_admission_tests {
-    use super::*;
-    use crate::recovery::gate::RecoveryGateConfig;
-
-    /// #322: after `Failed { retry_at = past }`, concurrent callers must
-    /// see **exactly one** `Probe` ticket, not a stampede. The CAS-based
-    /// single-probe claim lives in `admit_through_gate`. Each spawned
-    /// task parks on a `Barrier` before calling so the 32 attempts
-    /// really contend, instead of being serviced one at a time.
-    #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-    async fn expired_failed_state_admits_only_one_probe() {
-        let gate = Arc::new(RecoveryGate::new(RecoveryGateConfig {
-            max_attempts: 16,
-            base_backoff: Duration::from_millis(5),
-        }));
-
-        // Drive the gate into Failed { retry_at ≈ past }.
-        let ticket = gate.try_begin().expect("first ticket");
-        ticket.fail_transient("seed");
-        tokio::time::sleep(Duration::from_millis(20)).await;
-
-        async fn contend(
-            gate: Arc<RecoveryGate>,
-            barrier: Arc<tokio::sync::Barrier>,
-        ) -> (u32, u32) {
-            // Park here until every task is ready so we really stress
-            // the CAS claim.
-            barrier.wait().await;
-            let some_gate: Option<Arc<RecoveryGate>> = Some(gate);
-            match admit_through_gate(&some_gate) {
-                Ok(GateAdmission::Probe(ticket)) => {
-                    // Hold the probe until the test is done counting so
-                    // a second caller can't race in after a fast
-                    // resolve/fail cycle.
-                    tokio::time::sleep(Duration::from_millis(10)).await;
-                    drop(ticket);
-                    (1, 0)
-                },
-                Ok(GateAdmission::Open | GateAdmission::OpenGated(_)) => (0, 1),
-                Err(_) => (0, 1),
-            }
-        }
-
-        let barrier = Arc::new(tokio::sync::Barrier::new(32));
-        let mut handles = Vec::with_capacity(32);
-        for _ in 0..32 {
-            handles.push(tokio::spawn(contend(
-                Arc::clone(&gate),
-                Arc::clone(&barrier),
-            )));
-        }
-
-        let mut probes = 0u32;
-        let mut blocked = 0u32;
-        for h in handles {
-            let (p, b) = h.await.expect("admission task");
-            probes += p;
-            blocked += b;
-        }
-
-        assert_eq!(probes, 1, "exactly one Probe ticket must be granted (#322)");
-        assert_eq!(
-            probes + blocked,
-            32,
-            "every call must be accounted for: probes={probes}, blocked={blocked}",
-        );
-    }
-}
-
-#[cfg(test)]
-mod drain_race_tests {
-    use super::*;
-
-    /// Regression for the drain-race bug: previously `wait_for_drain`
-    /// did `tracker.1.notified().await` without pre-registering the
-    /// `Notified` future, so a handle dropping (and firing
-    /// `notify_waiters()`) in the window between the outer
-    /// `active == 0` check and the first `notified().await` poll would
-    /// leak the wakeup. Stall persisted until the full `drain_timeout`
-    /// elapsed.
-    ///
-    /// The fix pre-enables the `Notified` future and re-checks the
-    /// counter *after* registration, so a drop that completes the drain
-    /// mid-race is observed on the re-check and returns immediately.
-    ///
-    /// This test exercises the normal "handle drops while we're waiting"
-    /// path and asserts we return far sooner than the timeout.
-    #[tokio::test]
-    async fn wait_for_drain_returns_promptly_when_handle_drops() {
-        let mgr = Manager::new();
-        // Simulate one active handle.
-        mgr.drain_tracker.0.fetch_add(1, AtomicOrdering::Release);
-
-        let tracker = mgr.drain_tracker.clone();
-        tokio::spawn(async move {
-            tokio::time::sleep(Duration::from_millis(20)).await;
-            if tracker.0.fetch_sub(1, AtomicOrdering::Release) == 1 {
-                tracker.1.notify_waiters();
-            }
-        });
-
-        let start = Instant::now();
-        mgr.wait_for_drain(Duration::from_secs(30))
-            .await
-            .expect("handle drop must drain under the timeout");
-        let elapsed = start.elapsed();
-
-        assert!(
-            elapsed < Duration::from_secs(1),
-            "wait_for_drain should return within 1s when a handle drops, took {elapsed:?}"
-        );
-        assert_eq!(mgr.drain_tracker.0.load(AtomicOrdering::Acquire), 0);
-    }
-
-    /// Regression: if the counter reaches 0 *before* `wait_for_drain`
-    /// gets to pre-register the `Notified`, the post-enable re-check
-    /// must catch it and return immediately rather than stalling.
-    ///
-    /// We simulate the race by setting `active = 1` (so the outer
-    /// early-return doesn't fire), then immediately decrementing to 0
-    /// before `wait_for_drain` is polled.
-    #[tokio::test]
-    async fn wait_for_drain_catches_drop_via_recheck() {
-        let mgr = Manager::new();
-        mgr.drain_tracker.0.fetch_add(1, AtomicOrdering::Release);
-
-        // Decrement + notify synchronously — the counter is 0 before
-        // `wait_for_drain` is even called, but we want to prove that
-        // even if the outer check observed `active == 1` and then
-        // the counter hit 0 *between* that check and the inner enable,
-        // the inner re-check would catch it.
-        //
-        // Simulated here by priming the state and then calling
-        // wait_for_drain directly; the inner loop's re-check should
-        // fire on the very first iteration because the counter is
-        // already 0. The outer check is bypassed by the fetch_add
-        // above leaving active == 1 until... wait, we need to
-        // decrement BETWEEN the outer check and the inner enable.
-        //
-        // Easiest approximation: skip the outer early-return by
-        // keeping active = 1 through the outer check, then decrement
-        // via a spawned task that runs before wait_for_drain gets
-        // scheduler time.
-        let tracker = mgr.drain_tracker.clone();
-        tokio::task::yield_now().await;
-        let handle = tokio::spawn(async move {
-            // Yield so that wait_for_drain's outer load sees active = 1,
-            // then decrement before the inner poll happens.
-            tokio::task::yield_now().await;
-            if tracker.0.fetch_sub(1, AtomicOrdering::Release) == 1 {
-                tracker.1.notify_waiters();
-            }
-        });
-
-        let start = Instant::now();
-        mgr.wait_for_drain(Duration::from_secs(30))
-            .await
-            .expect("recheck path must drain under the timeout");
-        let elapsed = start.elapsed();
-        handle.await.unwrap();
-
-        assert!(
-            elapsed < Duration::from_secs(1),
-            "wait_for_drain must return promptly even under race, took {elapsed:?}"
-        );
-    }
-
-    /// #302: Abort policy must return a typed `DrainTimeout` error and
-    /// leave the registry untouched. Before the policy split
-    /// `graceful_shutdown` would log a warning and proceed to
-    /// `registry.clear()` anyway, turning a cooperative shutdown into a
-    /// logical use-after-free.
-    #[tokio::test]
-    async fn graceful_shutdown_abort_policy_returns_drain_timeout_error() {
-        let mgr = Manager::new();
-        // Simulate an outstanding handle.
-        mgr.drain_tracker.0.fetch_add(1, AtomicOrdering::Release);
-
-        let cfg = ShutdownConfig::default()
-            .with_drain_timeout(Duration::from_millis(50))
-            .with_drain_timeout_policy(DrainTimeoutPolicy::Abort);
-
-        let err = mgr
-            .graceful_shutdown(cfg)
-            .await
-            .expect_err("Abort policy must surface drain timeout");
-        match err {
-            ShutdownError::DrainTimeout { outstanding } => {
-                assert_eq!(outstanding, 1, "outstanding count mismatch");
-            },
-            other => panic!("wrong error variant: {other:?}"),
-        }
-    }
-
-    /// #302: Force policy must clear the registry and report the
-    /// outstanding-handle count in `ShutdownReport` so operators can see
-    /// exactly how much in-flight work was abandoned.
-    #[tokio::test]
-    async fn graceful_shutdown_force_policy_clears_registry_with_outstanding_count() {
-        let mgr = Manager::new();
-        mgr.drain_tracker.0.fetch_add(2, AtomicOrdering::Release);
-
-        let cfg = ShutdownConfig::default()
-            .with_drain_timeout(Duration::from_millis(50))
-            .with_drain_timeout_policy(DrainTimeoutPolicy::Force);
-
-        let report = mgr
-            .graceful_shutdown(cfg)
-            .await
-            .expect("Force policy must succeed");
-        assert!(report.registry_cleared);
-        assert_eq!(report.outstanding_handles_after_drain, 2);
     }
 }
